+++ lib/model/common/src/pool/spec.ts
@@ -8,6 +8,12 @@ export interface PObjectSpec {
   /** PObject kind discriminator */
   readonly kind: string;
 
+  /** Name is common part of PObject identity */
+  readonly name: string;
+
+  /** Domain is a set of key-value pairs that can be used to identify the object */
+  readonly domain?: Record<string, string>;
+
   /** Additional information attached to the object */
   readonly annotations?: Record<string, string>;
 }
@@ -87,6 +93,8 @@ export function extractAllColumns<D>(entry: JoinEntry<PColumn<D>>): PColumn<D>[]
         return;
       case 'slicedColumn':
         columns.set(entry.column.id, entry.column);
+        return;
+      case 'inlineColumn':
         return;
       case 'full':
       case 'inner':