+++ lib/ui/uikit/src/components/PlCheckbox/PlCheckboxBase.vue
@@ -1,23 +1,26 @@
 <script lang="ts" setup>
 import './pl-checkbox-base.scss';
-import { computed } from 'vue';
 
 defineEmits(['update:modelValue']);
 
 const props = defineProps<{
   modelValue: boolean;
   disabled?: boolean;
+  indeterminate?: boolean;
 }>();
-
-const checked = computed(() => props.modelValue);
 </script>
 
 <template>
   <div
     tabindex="0"
     class="pl-checkbox-base"
-    :class="{ checked, disabled }"
+    :class="{
+      ...(props.indeterminate
+        ? { indeterminate: true }
+        : { checked: props.modelValue }),
+      disabled,
+    }"
     @click="$emit('update:modelValue', !modelValue)"
-    @keydown.enter="$emit('update:modelValue', !modelValue)"
+    @keydown.space.prevent="$emit('update:modelValue', !modelValue)"
   />
 </template>