+++ tools/package-builder/src/commands/publish/packages.ts
@@ -27,7 +27,7 @@ export default class Packages extends Command {
     const { flags } = await this.parse(Packages);
     const logger = util.createLogger(flags['log-level']);
 
-    const core = new Core(logger);
+    const core = new Core(logger, { packageRoot: flags['package-root'] });
     core.pkg.version = flags.version;
     core.targetPlatform = flags.platform as util.PlatformType;
     core.allPlatforms = flags['all-platforms'];