+++ etc/blocks/pool-explorer/model/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @milaboratories/milaboratories.pool-explorer.model
 
+## 1.0.39
+
+### Patch Changes
+
+- Updated dependencies [4b685d6]
+  - @platforma-sdk/model@1.30.26
+
+## 1.0.38
+
+### Patch Changes
+
+- Updated dependencies [b0b80af]
+  - @platforma-sdk/model@1.30.24
+
+## 1.0.37
+
+### Patch Changes
+
+- Updated dependencies [ec15bb1]
+  - @platforma-sdk/model@1.30.21
+
 ## 1.0.36
 
 ### Patch Changes