+++ etc/blocks/ui-examples/block/CHANGELOG.md
@@ -1,5 +1,59 @@
 # @milaboratories/milaboratories.ui-examples
 
+## 1.0.256
+
+### Patch Changes
+
+- Updated dependencies [4b685d6]
+  - @platforma-sdk/model@1.30.26
+  - @milaboratories/milaboratories.ui-examples.model@1.1.57
+  - @milaboratories/milaboratories.ui-examples.ui@1.3.81
+
+## 1.0.255
+
+### Patch Changes
+
+- Updated dependencies [b0b80af]
+  - @platforma-sdk/model@1.30.24
+  - @milaboratories/milaboratories.ui-examples.model@1.1.56
+  - @milaboratories/milaboratories.ui-examples.ui@1.3.80
+
+## 1.0.254
+
+### Patch Changes
+
+- Updated dependencies [2aa0f3c]
+  - @milaboratories/milaboratories.ui-examples.ui@1.3.79
+
+## 1.0.253
+
+### Patch Changes
+
+- Updated dependencies [93d4a72]
+  - @milaboratories/milaboratories.ui-examples.ui@1.3.78
+
+## 1.0.252
+
+### Patch Changes
+
+- Updated dependencies [ec15bb1]
+  - @platforma-sdk/model@1.30.21
+  - @milaboratories/milaboratories.ui-examples.model@1.1.55
+  - @milaboratories/milaboratories.ui-examples.ui@1.3.77
+
+## 1.0.251
+
+### Patch Changes
+
+- Updated dependencies [d82a569]
+  - @milaboratories/milaboratories.ui-examples.ui@1.3.76
+
+## 1.0.250
+
+### Patch Changes
+
+- @milaboratories/milaboratories.ui-examples.workflow@1.1.11
+
 ## 1.0.249
 
 ### Patch Changes