+++ lib/node/pl-drivers/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @milaboratories/pl-drivers
 
+## 1.5.57
+
+### Patch Changes
+
+- Updated dependencies [b0b80af]
+  - @milaboratories/pl-model-common@1.15.0
+
+## 1.5.56
+
+### Patch Changes
+
+- Updated dependencies [94468e0]
+  - @milaboratories/pl-model-common@1.14.1
+
+## 1.5.55
+
+### Patch Changes
+
+- 6ab1be8: updated check network
+
+## 1.5.54
+
+### Patch Changes
+
+- b8c7847: extract download_blob to a separate folder; jest -> vitest
+
 ## 1.5.53
 
 ### Patch Changes