+++ sdk/model/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @platforma-sdk/model
 
+## 1.30.26
+
+### Patch Changes
+
+- 4b685d6: - forceTraceElements for deriveLabels
+  - exclude option for multiple methods in anchoredId API
+
+## 1.30.24
+
+### Patch Changes
+
+- b0b80af: - fix for label derivation with enough diversity solaly in labels
+  - fallback behaviour for derive labels function
+  - deduplication of columns in column collection
+- Updated dependencies [b0b80af]
+  - @milaboratories/pl-model-common@1.15.0
+
+## 1.30.21
+
+### Patch Changes
+
+- ec15bb1: createPFrameForGraphs support for DataInfo<TreeNodeAccessor>
+- Updated dependencies [94468e0]
+  - @milaboratories/pl-model-common@1.14.1
+
 ## 1.30.11
 
 ### Patch Changes