+++ tools/block-tools/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @platforma-sdk/block-tools
 
+## 2.5.47
+
+### Patch Changes
+
+- @milaboratories/pl-model-middle-layer@1.7.32
+
+## 2.5.46
+
+### Patch Changes
+
+- Updated dependencies [94468e0]
+  - @milaboratories/pl-model-middle-layer@1.7.31
+
+## 2.5.45
+
+### Patch Changes
+
+- Updated dependencies [8e23a2e]
+  - @milaboratories/pl-model-middle-layer@1.7.30
+
 ## 2.5.44
 
 ### Patch Changes